allows to disassemble elf files without sections (#1085)
Usually, when an ELF binary contains both segments and sections,
we use the latter to determine code regions. Also, sections are
used to print disassembled output via bap <file> -dasm.

And if there are no sections in a binary, we can do nothing with
such file.
This PR resolves this issue and provide all the necessary
information from segments and allow us to process such kind of files.