chains tasks into trees
Each task now puts the parent task into its constraints.

Note, it would probably be more optimal to to store the models, but I
am not sure that it would be correct. On the first thought, if we will
inherit the model from the parent task we will significantly limit the
solution space.