fixes BIl.eval memory operations for different sizes of values
Bil.Eval still relies on the assumption that memory can address only values of 1 byte length. And once the stored/loaded value size
differs from memory value size, we try to simplify an expression. But the simplification in its turn can leave an expression unchanged
since it takes into account both memory value size and size of a stored/loaded value. And that's how we get an infinite loop: since an expression is unchanged, we try to eval it again, simplify, and so on.
This PR fixes this problem and checks both the size of stored/loaded value and memory value size.