prunes unreachable code in the optimization pass (#1095)
* Prunes unreachable code in the optimization pass

This PR fixes #1092 :
The optimization pass may make some code unreachable, propagating a constant value to a branch guard.

For example, some code can become unreachable due to jumps that are
never taken because of condition:
```
0000046a: when 0 goto %0000045b
```

Or due to the previous jump that is taken always:
```
00000471:
00000473: goto %00000451
00000474: goto %00000455
```

This PR fixes this problem and removes blocks that are unreachable due to infeasible jumps and these jumps as well.

Also, this PR fixes the computation of digest in the optimization pass. Previously, digest for each subroutine contained tids, that is not quite right, because a program can have different tids on each bap execution, and actually it happens quite often. And it leads to the cache explosion since instead of loading from cache we store some duplicated data that are different only by the tids values. This PR solves this problem and uses addresses instead.

* removed old code

* renaming

* updates digesting

* minor updates