adds the new Memory.add_region function
It is the same as the old `allocate` (which is still available, no
worries), but takes two words instead of one word and int, which makes
it possible to create regions whose size do not fit into OCaml int,
which is quite possible during execution in a randomized mode.

And makes the code much eaiser to write as now we don't need to deal
with conversion errors.