use the path as the package not the digest
Rationale: in a multi-project analysis we will need to know the
package to obtain the information relative to a project and getting
the digest is a bit long road. Moreover, it is less
readable. Concerning the possible conflicts from the different files
with the same name, it is quite unlikely, since the caching mechanism
separates knowledge bases by the digests and other systems that we or
anyone else will develop will manage their own separation schemes.