fixes an inifinite loop in certain Primus Lisp analysis
Since now Primus Lisp operations are expressed in terms of the Primus
interpreter (in order to enable analysis of the code that is stubbed
in Primus Lisp), we have the `eval-cond` signal sent every time we
make any comparison operation in Lisp, e.g., `if`, `when`, `or`,
`and`, etc. Therefore, any method for `eval-cond` that does any
comparison operation (branching) will stall in an infinite loop.

A workaround is to use another observation that occurs only on jumps
that are situated in the binary code, but in general we shall address
it and make sure that `eval-cond` method is usable.