switches to a non-alternate llvm, removes the LLVM_CONFIG step
First of all, `/usr/local/opt/llvm/bin/llvm-config`
is llvm-config of LLVM that is shipped with MacOS and which is
stripped from any non-native targets so we can't use it. The
requested llvm@9 is installed at `/usr/local/opt/llvm@9` and is not
symlinked to `/usr/local` because it is keg-only:

```
llvm@9 is keg-only, which means it was not symlinked into /usr/local,
because this is an alternate version of another formula.

If you need to have llvm@9 first in your PATH run:
  echo 'export PATH="/usr/local/opt/llvm@9/bin:$PATH"' >> /Users/runner/.bash_profile

For compilers to find llvm@9 you may need to set:
  export LDFLAGS="-L/usr/local/opt/llvm@9/lib"
  export CPPFLAGS="-I/usr/local/opt/llvm@9/include"
```

llvm@10 aka llvm is also keg-only so it will be installed only into
the cellar. I have to guess whether the llvm-config binary will be
named llvm-config or llvm-config-9, but I bet for the former.