Try to use GitHub Actions (#1115)
* Try to use GitHub Actions

* Fix LLVM installation for Homebrew

* Disable install for now

* MacOS fixes

* Disable Windows for now

* Fix CI

* fixes CI, build on ubuntu only

* launch unit tests after bap installed

* try set cache home dir via env

* wip

* wip

* Grammar eye-candy.

* Try MacOS again

* switches to a non-alternate llvm, removes the LLVM_CONFIG step

First of all, `/usr/local/opt/llvm/bin/llvm-config`
is llvm-config of LLVM that is shipped with MacOS and which is
stripped from any non-native targets so we can't use it. The
requested llvm@9 is installed at `/usr/local/opt/llvm@9` and is not
symlinked to `/usr/local` because it is keg-only:

```
llvm@9 is keg-only, which means it was not symlinked into /usr/local,
because this is an alternate version of another formula.

If you need to have llvm@9 first in your PATH run:
  echo 'export PATH="/usr/local/opt/llvm@9/bin:$PATH"' >> /Users/runner/.bash_profile

For compilers to find llvm@9 you may need to set:
  export LDFLAGS="-L/usr/local/opt/llvm@9/lib"
  export CPPFLAGS="-I/usr/local/opt/llvm@9/include"
```

llvm@10 aka llvm is also keg-only so it will be installed only into
the cellar. I have to guess whether the llvm-config binary will be
named llvm-config or llvm-config-9, but I bet for the former.

* sets TMPDIR env var to pass tests on mac os

* updates testsuite

* do not run on each push

The master branch is already restricted from pushes so the code can reach the master branch only through a pull requested. We don't want to the same work twice.

Also removes (the commented) windows line, we will return it back, once we will have time for that, see also #512

Co-authored-by: oleg <forown@yandex.ru>
Co-authored-by: ivg <ivg@ieee.org>