Tweak jumping and eval cond (#1142)
* tweaks eval-cond and jumping observations

Now the guard condition is computed from all outcoming edges, which
simplifies taint analysis and symbolic execution. This is true for
both jumping and eval-cond.

* simplifies and optimizes symbolic executor

now, when eval-cond takes into account path constraint (i.e., that
previous branches in the same basic block were not taken) we can
simplify the code of the symbolic executor that is responsible for
executing physical (i.e., represented with actual basic blocks)
braches as we now have to investigate only non-taken branches.

* removes the eval-cond signal from the Primus Lisp interface

We might put it back before the release, but right now it is the
easiest and less error-prone solution.

* bumps the timeout value in the symexec tests

Since functional tests take five times more time on the Mac OS X
machines, it is not surprising that we are occasionally hitting the
timeout on them, see also #1136. I hope that double would be enough.