introduces code units
A new class of Core Theory objects that represents compilation units,
libaries, executables, and other aggregations of programs.

The path property now belongs to the unit, not the the program itself,
so we basically introduced an extra layer of indirection.

One of the important properties of units is that they could be shifted
in memory, which paves the road to the llvm-base fix.