fixes several bugs introduced with the radare2 backend (#1159)
1) the names weren't properly demangled therefore all calls to plt
functions were looking as `imp.X` instead of `X`

2) all entries from the symbol table were interpreted as functions,
including data objects, references to data, and relocations, as a
result we had a few false function starts and bogus functions such as
$a, $b, etc

3) radare2 is providing names even if the name is not known to it,
suggesting bogus `func.X` names. This will prevent other symbolizers
from providing their own names and induce radare2 default function
naming on us, which we don't really want (we prefer to stick with
sub_XXXX for subroutines with unknown name).

Unfortunately, our testing infrastructure were unable to detect this
issues because all tests are using pre-recorded symbolic information,
thus symbolizers are basically not covered. I think that it is
reminiscent of the old times when we were using Travis and unable to
bring objdump and radare2 on it. We can now try to remove as much
static symbols as possible.