fixes barrier instructions with delay slots
During disassembling we are processing instructions in linear order
not execution and to detect if there is a fallthrough we were looking
at the `barrier` property of the last disassembled instruction, which,
in case of MIPS and other architectures with delay slots, wasn't even
a jump.

Instead, we should look into the disassembler state in which we have
recording for each barrier and which is even faster.