publishes and documents the new disassembler engine
The engine was introduced in BAP 2.0 but wasn't fully published or
documented.

Some parts of the interface were published but were lacking the
documentation, some were just missing.

We also publish the low-level interface to the disassembler engine,
which is totally new.

Last but not least we provide a precise description of the algorithm
and the interaction of the engine with the knowledge base.