enables biased information sources
We assumed by default that all our information sources (rooters,
symbolizers, and branchers) are unbiased but it wasn't true for
the sources that we created of images (using corresponding of_image)
that were already operating using the biased information from the
loader.

To fix this issue we added a hidden parameter to mark information
sources as biased or unbiased and perform bias substraction (and also
addition in case when the destinations are provided) based on this
parameter. We may later make it public, but so far it is only set for
information sources that are derived from images.

Also the common code between information sources were factored out
into the Bap_disasm_source module.