enables unallocated memories in symbolic executor (#1177)
When the symbolic executor tries to initialize a memory location that is not yet
mapped it fails with an error, when the location belongs to a memory
that wasn't set as symbolic (which is most likely just the default
virtual memory). Instead of failing, we shall just intialize the
memory location and let the pagefault handler of the memory randomizer
to map it for us.