expects fail in the macho loader (#1181)
Our nightly tests fail on the macho loader tests with ko/powerpc-xnu-kext file. The reason is in the small hack in ogre loader, which works for ELF files but does not work for the other ones.

When we deal with ELF relocatable files, we want to see the same addresses, as they are in the objdump, i.e. to have an address of the first symbol to be zero. And we subtract an address of the first section from the base address to reach it. But the same thing doesn't work for other file types, so we have to expect fails in the macho loaders test for some time, while this bug won't be fixed.