stores log files for functional test (#1202)
* stores log files for functional test

fixes https://github.com/BinaryAnalysisPlatform/bap/issues/1201

For every functional test, we create a folder in which we store log files. And in case of any fail, it will be easier to find a problem since we'll able to find a corresponded log file for the failed test.

Also, this PR fixes a minor bug in our Github Actions workflows: when we use `matrix`, we don't have multiple artifacts, depending
on matrix size. Instead, for every `upload-artifact` action we have only one artifact per workflow, i.e. artifacts are overridden.
This PR fixes this problem and names artifact according to the matrix element it is uploaded for, e.g. `log-ubuntu-4.07`

* fixes logdir