adds an action for release automatization
After trying different ready-to-use actions, I think I found the best one (so far).

The good things about this action are:
1) It allows us to upload many files using file patterns.
2) Generates changelog automatically.
3) Rewrites releases with the same tag, so there is no need somehow to query `github` and to delete the previous release.

The proposal workflow is to run this action once a week. When `bap` will be ready for the next release, just rename the current
tag from `v2.2.0-alpha` to `v2.2.0` and drop the `prerelease` checkbox.