reduces the default timeout back to 16 ms
It looks like that setting it to high makes the symbolic executor very
slow, as instead of discarding overly big formulas we spend more time
on solving it, rather than exploring other states. It is probably, why
our tests on symbolic executor are failing, though can't say for sure,
it is hard to bisect this issue, given that z3 timing is significantly
non-deterministic.